Disable OoaK alchemy until max rolls.
An ascension is essentially won once 5 MoaKs have been rolled as the players who have the MoaKs can alchemize an OoaK, split the dough, and get to max rolls. Some people strongly dislike this, as it feels cheap. This disallows OoaK alchemization until the player has achieved max rolls.
The rolls required does adjust with ascensions. It should also allow for more OoaKs to be added and will disallow creation of those, too. as long as the OoaK is in values.all_one_of_a_kind.